Fixing height bug for images with different aspect ratios
For the longest time there's been a very small height difference in the album view for images with different heights. This was caused by padding between the images and surprisingly hard to get rid of.

Because the image widths are calculated in percentage at a point where I don't know the page width, and because I don't want the space between images to be specified in percentages, I've had to create the space between the images by making the box aruond them smaller by the same amount as the padding, then add a negative margin of the same size and finally padding between the images. This way the percentage widths of the images can be output to the page and the height is rendered the right way.

I'm very pleased.