Fixes to image sizing when presented in albums
I've had an issue with blurry vertical images for a long time. When
investigating I found two issues were causing this:

1. Css style rules around media sizes means we only serve up images of
up to half the screen-width (except if they are marked as covers). I've
changed this to serve full size images up to a screen width of 800px and
roughly two-third sized images for larger screen widths (given that even
though images are displayed side by side, an image might still be next
to a vertical image with smaller width.

2. Vertical images are cropped by height and not width when uploaded.
This means that when we request an image of size 600x450, it'll only be
300x450 if it's vertical. To fix this I'm requesting double the image
size for vertical images. So for an image that needs to fit a space of
600x450, I'll request a vertical image of size 1280x980 (because this
hapepns to be the discrete size I've precomputed), which when then image
is vertical will return me an image of size roughly 650x980.