If last image is a cover, it should stay last.
Currently, if the last image is a cover image and the second to last
isn't, the two will swap places on the page and both will be rendered as
covers.

This is both a bit silly, and also means we don't load an image
big enough to the page because the css rules only consider the last
image _in the album_ as a cover sized image. As a consequence if the
last image in the album is already a cover and swaps place with the
second to last, it is put on the page in non-cover resolution.

As a fix, I've changed so that the last image will always be the last.
If it's a cover image, it will sometimes be downgraded to a normal
image, to make the amount of images fit the page.