Enabling video playback on ifany
It's done by inserting a video tag where we otherwise would insert an
image tag. It was remarkably straight forward. Most of the code is left
alone, which is why we're still referring to videos as 'images'. Videos
will have the `is_video` attribute set to `true`.

There's a few missing steps:
- Upload 'posters' for thumbnails (or just make so videos aren't used)
- See if we can make an image show up for a video that isn't loaded yet
on a phone