Use video thumbs for album covers, video thumbs and album thumbs
When a video is uploaded with photo-tools, we also upload the first
frame in the usual sizes as images. This means that if a video for
example is marked as a cover, we can still show a still from the video
as the album cover.

Another benefit is that we can set the `poster` attribute in the video
tag which lets phones download a placeholder image to show instead of
loading the video. Without it, the video just shows as a black frame
when browsing on a phone.

This commit also fixes the embarrasing issue that images that were not
published, were apperently still shown in albums.