Closing S3 inputstreams (oups)
I'm seeing out of memory exceptions thrown when I reload albums with a
lot of bigger images and I'm wondering if it might have to do with the
fact that I never close the inputstreams I get when fetching s3 objects.

I have been too lazy to reproduce the out of memory error in my local
dev environment, but I have checked that this change doesn't break
anything so without having confidence that it necessarily fixed the
issue, I'm going to push it under the belief that it will if nothing
less have improved the code a little bit.