Use html image srcsets instead of css viewports for image selection
Using srcsets
-------------

The album pages used to be set up using css to load the optimal image
sizes for the screen width by adding the image as a background-image to
a div and including a lot of viewport rules. This worked ok, but it
meant that all the logic for how big an image might be and what image to
fetch was done manually.

Using image srcsets I just specify the width of the available image
sizes (e.g. `srcset="img_1234_400x300.jpg 400w, img_1234_800x600.jpg
800w"`) as well as the size of the image in viewport percentage (e.g.
`sizes="(min-width: 800px) 56vw, 100vw"` -- Use 56% of screen width if
screen is >= 800px wide, otherwise use 100%). Based on these bits of
information, the browser picks the best image based on resolution and
dpi.

Exact widths
------------

For some reason I had never really bothered to compute the exact width
of vertical images. Due to not quite thought through resizing on
uploading, vertical images (bigger height than width) would be much
narrower than the constraints in the image name, so for example
`img_1234_400_300.jpg` might only be 255px wide if it was vertical
because it would have been resized to be within the boundary box of
400px by 300px.

Realising that they were smaller I had just kind of assumed that if the
aspect ratio was roughly "normal", the width would be approximately two
times smaller than the constraint box. In hindsight I know the original
image size and have all the information available to compute the exact
width of any image no matter the aspect ratio, so I've changed the code
to do that nicely.

Async Rework
------------

There's one final change, and I'm probably mentioning it here at the end
because it really ought to have been it's own commit and doesn't quite
work as I expected it to: Looking a bit closer at the
`plans/Gallery.scala`, I seem to have been mostly using blocking calls
together with a netty async plan which is a bit embarrassing in that
there's no point in having async request handlers if the calls just
block instead of actually spin off some work on a separate thread.

I've fixed this for getting files from s3 and it works and all, but it's
not really solving the issue that made me look at this code in the first
place, in that for larger files I still seem to put the server in an
awkward state of mind where it'll serve the first eight or ten images
and then just stop with no errors. I'm not sure what's going on.

I'm hoping to fix it by just serving the files straight from s3 instead.