Serving album images from S3
Instead of making the bucket public, I'm using presigned urls. This
turned out to be a bit of a faff because the urls generated for a bucket
with dots aren't valid for AWS' SSL certificarte. The official advice
from AWS seems to be that changing the bucket name to something not
involving dots is the way to go. That's a bit silly.

I found a workaround in that the s3 urls also accept omitting a bucket
and including the bucket name as part of the key. I'm slightly worried
this might be a legacy format with limited support, but for now it
neatly circumvents the issue