Add password to show non-public images
I've been wanting to use ifany for more private images alongside the
public albums I already host on the website. For this purpose I'm
interpreting the url part after the album as the album password and
checking it against a hash stored in the database:

```
www.ifany.org/<gallery>/<album>/<password>/
```

The feature work by sha-256 encoding the album-id concatenated with the
password and comparing it to the value found in the database. Then on
the album editing side of things, I'll have to add a password field
which will get sha-256 encoded as well and written to the database.

This solution is a compromise between several constraints:
- I want to be able to link other people to an album showing private
  pictures
- I didn't want to rely on security through obscurity (except maybe the
  obscurity of the passwords themselves)
- I wanted this change to involve the least amount of faff possible

I'm pretty pleased with this change. It involves very little change to
the source code for quite a substantial feature. Nice job, me.